Examples/WebApps/GlupWebApp/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
tensorflow==2.5.0
numpy==1.19.2
h5py==3.1.0
six==1.15.0
scipy==1.4.1
Pillow==8.2.0
opencv-python-headless==4.5.3.56
